fix: renaming & adjust entity
