+++ in_a_bottle/lib/common/widget/locked/lock_widget.dart
@@ -3,7 +3,7 @@ import 'package:in_a_bottle/_shared/widgets/widget_factory/crud_widget.dart';
 import 'package:in_a_bottle/_shared/widgets/widget_factory/form_factory.dart';
 import 'package:in_a_bottle/_shared/widgets/widget_factory/widget_text_field.dart';
 import 'package:in_a_bottle/common/widget/locked/lock_bloc.dart';
-import 'package:in_a_bottle/local_message/local/local_dto.dart';
+import 'package:in_a_bottle/local_message/local/local.dart';
 import 'package:fancy_stream/fancy_stream.dart';
 
 class LockWidget extends StatefulWidget {
