+++ in_a_bottle/lib/local_message/direct_message/widgets/create/direct_message_widget.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:in_a_bottle/_shared/widgets/widget_factory/crud_widget.dart';
-import 'package:in_a_bottle/local_message/message/widgets/create/direct_message_bloc.dart';
+import 'package:in_a_bottle/local_message/direct_message/widgets/create/direct_message_bloc.dart';
 import 'package:fancy_stream/fancy_stream.dart';
 import 'package:in_a_bottle/_shared/widgets/widget_factory/widget_text_field.dart';
