+++ in_a_bottle/lib/local_message/direct_message/widgets/interact/interact_direct_message_bloc.dart
@@ -1,7 +1,7 @@
 import 'package:fancy_stream/fancy_stream.dart';
-import 'package:in_a_bottle/local_message/message/message_model.dart';
-import 'package:in_a_bottle/local_message/message/message_repository.dart';
-import 'package:in_a_bottle/local_message/message/widgets/interact/interact_direct_message_event.dart';
+import 'package:in_a_bottle/local_message/direct_message/direct_message.dart';
+import 'package:in_a_bottle/local_message/direct_message/direct_message_repository.dart';
+import 'package:in_a_bottle/local_message/direct_message/widgets/interact/interact_direct_message_event.dart';
 import 'package:in_a_bottle/local_message/reaction/type_reaction.dart';
 import 'package:in_a_bottle/local_message/reaction/user_reaction.dart';
 import 'package:in_a_bottle/session/session_repository.dart';
@@ -11,7 +11,7 @@ import 'package:in_a_bottle/_shared/archtecture/base_bloc.dart';
 
 class InteractDirectMessageBloc extends BaseBloc<InteractDirectMessageEvent> {
   static const String route = '/interactDirectMessage';
-  final MessageRepository messageRepository;
+  final DirectMessageRepository messageRepository;
   final SessionRepository sessionRepository;
 
   InteractDirectMessageBloc(
@@ -32,36 +32,37 @@ class InteractDirectMessageBloc extends BaseBloc<InteractDirectMessageEvent> {
   }
 
   Future<void> _applyReaction(
-      TypeReaction reaction, Message message) async {
+      TypeReaction reaction, DirectMessage directMessage) async {
     final entity = await _buildEntity();
 
     final session = await sessionRepository.load();
     final reactionOfUser =
         UserReaction(createdBy: session.user, reaction: reaction);
 
-    bool exist = message.reactions.contains(reactionOfUser);
+    bool exist = directMessage.message.reactions.contains(reactionOfUser);
 
     if (exist) {
-      message.reactions.remove(reactionOfUser);
+      directMessage.message.reactions.remove(reactionOfUser);
     } else {
-      message.reactions.add(reactionOfUser);
+      directMessage.message.reactions.add(reactionOfUser);
     }
 
     await messageRepository.save(entity);
 
-    final hubReactionsCounted = await messageRepository.loadByKey(entity.selector);
+    final hubReactionsCounted =
+        await messageRepository.loadByKey(entity.selector);
 
     _updateHubOnScreen(hubReactionsCounted);
   }
 
-  void _updateHubOnScreen(Message entity) {
-    dispatchOn<Message>(entity, key: DirectMessageForm.directMessage);
+  void _updateHubOnScreen(DirectMessage entity) {
+    dispatchOn<DirectMessage>(entity, key: DirectMessageForm.directMessage);
   }
 
-  Future<Message> _buildEntity() async {
+  Future<DirectMessage> _buildEntity() async {
     final map = valuesToMap<DirectMessageForm>();
 
-    return map[DirectMessageForm.directMessage] as Message;
+    return map[DirectMessageForm.directMessage] as DirectMessage;
   }
 }
