+++ in_a_bottle/lib/local_message/direct_message/widgets/interact/interact_direct_message_event.dart
@@ -1,4 +1,4 @@
-import 'package:in_a_bottle/local_message/message/message_model.dart';
+import 'package:in_a_bottle/local_message/direct_message/direct_message.dart';
 import 'package:in_a_bottle/local_message/reaction/type_reaction.dart';
 
 abstract class InteractDirectMessageEvent {}
@@ -11,7 +11,7 @@ class LoadDirectMessage extends InteractDirectMessageEvent {
 
 class SelectReaction extends InteractDirectMessageEvent {
   final TypeReaction reaction;
-  final Message message;
+  final DirectMessage message;
 
   SelectReaction(this.reaction, this.message);
 }
