+++ in_a_bottle/lib/local_message/direct_message/widgets/interact/interact_direct_message_widget.dart
@@ -2,9 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:in_a_bottle/_shared/injection/injector.dart';
 import 'package:in_a_bottle/common/widget/locked/lock_widget.dart';
 import 'package:fancy_stream/fancy_stream.dart';
-import 'package:in_a_bottle/local_message/message/message_model.dart';
-import 'package:in_a_bottle/local_message/message/widgets/interact/interact_direct_message_bloc.dart';
-import 'package:in_a_bottle/local_message/message/widgets/interact/interact_direct_message_event.dart';
+import 'package:in_a_bottle/local_message/direct_message/direct_message.dart';
+import 'package:in_a_bottle/local_message/direct_message/widgets/interact/interact_direct_message_bloc.dart';
+import 'package:in_a_bottle/local_message/direct_message/widgets/interact/interact_direct_message_event.dart';
 import 'package:in_a_bottle/local_message/reaction/reaction_widget.dart';
 
 class InteractDirectMessageWidget extends StatefulWidget {
@@ -37,9 +37,9 @@ class _InteractDirectMessageWidgetState
   Widget build(BuildContext context) {
     return Scaffold(
       body: Container(
-          child: StreamBuilder<Message>(
-              stream: _bloc.streamOf<Message>(
-                  key: DirectMessageForm.directMessage),
+          child: StreamBuilder<DirectMessage>(
+              stream:
+                  _bloc.streamOf<DirectMessage>(key: DirectMessageForm.directMessage),
               builder: (context, snpashot) {
                 final dm = snpashot.data;
                 return LockWidget(
@@ -50,7 +50,7 @@ class _InteractDirectMessageWidgetState
                       onReactionChange: (r) =>
                           _bloc.dispatchOn<InteractDirectMessageEvent>(
                               SelectReaction(r, dm)),
-                      userReactions: dm?.reactions,
+                      userReactions: dm?.message?.reactions,
                     )
                   ]),
                 );
