+++ in_a_bottle/lib/local_message/hub/create/create_hub_message_widget.dart
@@ -1,18 +1,18 @@
 import 'package:flutter/material.dart';
 import 'package:in_a_bottle/_shared/widgets/widget_factory/crud_widget.dart';
 import 'package:in_a_bottle/_shared/widgets/widget_factory/widget_text_field.dart';
-import 'package:in_a_bottle/local_message/hub/create/create_chat_bloc.dart';
+import 'package:in_a_bottle/local_message/hub/create/create_hub_message_bloc.dart';
 import 'package:fancy_stream/fancy_stream.dart';
 
-class CreateChatWidget extends StatefulWidget {
+class CreateHubMessageWidget extends StatefulWidget {
   @override
-  _CreateChatWidgetState createState() => _CreateChatWidgetState();
+  _CreateHubMessageWidgetState createState() => _CreateHubMessageWidgetState();
 }
 
-class _CreateChatWidgetState extends State<CreateChatWidget> {
+class _CreateHubMessageWidgetState extends State<CreateHubMessageWidget> {
   @override
   Widget build(BuildContext context) {
-    return CrudWidget<ChatForm, ChatError, CreateChatBloc>(
+    return CrudWidget<ChatForm, ChatError, CreateHubMessageBloc>(
         builder: (_bloc, _factory) {
       return Container(
         child: SingleChildScrollView(
