+++ in_a_bottle/lib/local_message/hub/hub_message.dart
@@ -1,22 +1,22 @@
 import 'dart:convert';
 
 import 'package:equatable/equatable.dart';
-import 'package:in_a_bottle/_shared/archtecture/base_model.dart';
 
+import 'package:in_a_bottle/_shared/archtecture/base_model.dart';
 import 'package:in_a_bottle/home/home_feed.dart';
-import 'package:in_a_bottle/local_message/hub/chat_message.dart';
-import 'package:in_a_bottle/local_message/local/local_dto.dart';
+import 'package:in_a_bottle/local_message/local/local.dart';
+import 'package:in_a_bottle/local_message/message/message.dart';
 import 'package:in_a_bottle/user/user_dto.dart';
 
-class Chat extends Equatable implements BaseModel, HomeFeed {
+class HubMessage extends Equatable implements BaseModel, HomeFeed {
   final String selector;
   final Local local;
   final List<User> admin;
   final String title;
   final User createdBy;
-  final List<ChatMessage> messageChat;
+  final List<Message> messageChat;
 
-  Chat({
+  HubMessage({
     this.selector,
     this.local,
     this.admin,
@@ -25,15 +25,15 @@ class Chat extends Equatable implements BaseModel, HomeFeed {
     this.messageChat,
   });
 
-  Chat copyWith({
+  HubMessage copyWith({
     String selector,
     Local local,
     List<User> admin,
     String title,
     User createdBy,
-    List<ChatMessage> messageChat,
+    List<Message> messageChat,
   }) {
-    return Chat(
+    return HubMessage(
       selector: selector ?? this.selector,
       local: local ?? this.local,
       admin: admin ?? this.admin,
@@ -44,7 +44,7 @@ class Chat extends Equatable implements BaseModel, HomeFeed {
   }
 
   Map<String, dynamic> toMap() {
-    return <String, dynamic>{
+    return {
       'selector': selector,
       'local': local?.toMap(),
       'admin': admin?.map((x) => x?.toMap())?.toList(),
@@ -54,26 +54,22 @@ class Chat extends Equatable implements BaseModel, HomeFeed {
     };
   }
 
-  factory Chat.fromMap(Map<String, dynamic> map) {
+  factory HubMessage.fromMap(Map<String, dynamic> map) {
     if (map == null) return null;
-
-    return Chat(
-      selector: map['selector'] as String,
-      local: Local.fromMap(map['local'] as Map<String, dynamic>),
-      admin: List<User>.from((map['admin'] as Iterable<Map<String, dynamic>>)
-          ?.map<User>((x) => User.fromMap(x))),
-      title: map['title'] as String,
-      createdBy: User.fromMap(map['createdBy'] as Map<String, dynamic>),
-      messageChat: List<ChatMessage>.from(
-          (map['messageChat'] as Iterable<Map<String, dynamic>>)
-              ?.map<ChatMessage>((x) => ChatMessage.fromMap(x))),
+  
+    return HubMessage(
+      selector: map['selector'],
+      local: Local.fromMap(map['local']),
+      admin: List<User>.from(map['admin']?.map((x) => User.fromMap(x))),
+      title: map['title'],
+      createdBy: User.fromMap(map['createdBy']),
+      messageChat: List<Message>.from(map['messageChat']?.map((x) => Message.fromMap(x))),
     );
   }
 
   String toJson() => json.encode(toMap());
 
-  factory Chat.fromJson(String source) =>
-      Chat.fromMap(json.decode(source) as Map<String, dynamic>);
+  factory HubMessage.fromJson(String source) => HubMessage.fromMap(json.decode(source));
 
   @override
   bool get stringify => true;
