+++ in_a_bottle/lib/local_message/hub/interact/interact_hub_message_event.dart
@@ -1,0 +1,17 @@
+import 'package:in_a_bottle/local_message/message/message.dart';
+import 'package:in_a_bottle/local_message/reaction/type_reaction.dart';
+
+abstract class InteractHubMessageEvent {}
+
+class LoadChat extends InteractHubMessageEvent {
+  final String selector;
+
+  LoadChat(this.selector);
+}
+
+class SelectReaction extends InteractHubMessageEvent {
+  final TypeReaction reaction;
+  final Message messageChat;
+
+  SelectReaction(this.reaction, this.messageChat);
+}
