+++ in_a_bottle/lib/local_message/talk/widget/create/talk_bloc.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:in_a_bottle/_shared/archtecture/crud_bloc.dart';
 import 'package:in_a_bottle/_shared/location/location_repository.dart';
 import 'package:in_a_bottle/_shared/route/navigator.dart' as nav;
-import 'package:in_a_bottle/local_message/local/local_dto.dart';
-import 'package:in_a_bottle/local_message/talk/talk_model.dart';
+import 'package:in_a_bottle/local_message/local/local.dart';
+import 'package:in_a_bottle/local_message/talk/talk.dart';
 import 'package:in_a_bottle/local_message/talk/talk_repository.dart';
 import 'package:meta/meta.dart';
 import 'package:fancy_stream/fancy_stream.dart';
