+++ in_a_bottle/lib/local_message/talk/widget/interact/interact_talk_event.dart
@@ -1,6 +1,5 @@
-import 'package:in_a_bottle/local_message/message/message_model.dart';
+import 'package:in_a_bottle/local_message/message/message.dart';
 import 'package:in_a_bottle/local_message/reaction/type_reaction.dart';
-import 'package:in_a_bottle/local_message/talk/talk.dart';
 
 abstract class InteractTalkEvent{
