+++ in_a_bottle/lib/local_message/talk/widget/interact/interact_talk_widget.dart
@@ -3,10 +3,10 @@ import 'package:in_a_bottle/_shared/injection/injector.dart';
 import 'package:in_a_bottle/_shared/widgets/widget_factory/form_factory.dart';
 import 'package:in_a_bottle/common/widget/locked/lock_widget.dart';
 import 'package:fancy_stream/fancy_stream.dart';
-import 'package:in_a_bottle/local_message/message/message_model.dart';
+import 'package:in_a_bottle/local_message/direct_message/direct_message.dart';
+import 'package:in_a_bottle/local_message/message/message.dart';
 import 'package:in_a_bottle/local_message/reaction/reaction_widget.dart';
 import 'package:in_a_bottle/local_message/talk/talk.dart';
-import 'package:in_a_bottle/local_message/talk/talk_model.dart';
 import 'package:in_a_bottle/local_message/talk/widget/interact/interact_talk_bloc.dart';
 import 'package:in_a_bottle/local_message/talk/widget/interact/interact_talk_event.dart';
