+++ in_a_bottle/lib/main.dart
@@ -6,8 +6,8 @@ import 'package:in_a_bottle/home/home_di.dart';
 import 'package:in_a_bottle/home/widgets/home_widget.dart';
 import 'package:in_a_bottle/_shared/route/navigator.dart'
     as interface_navigator;
-import 'package:in_a_bottle/local_message/hub/chat_di.dart';
-import 'package:in_a_bottle/local_message/message/message_di.dart';
+import 'package:in_a_bottle/local_message/hub/hub_message_di.dart';
+import 'package:in_a_bottle/local_message/direct_message/direct_message_di.dart';
 import 'package:in_a_bottle/local_message/talk/talk_di.dart';
 import 'package:in_a_bottle/session/session_bloc.dart';
 import 'package:in_a_bottle/session/session_di.dart';
@@ -20,11 +20,11 @@ import 'package:flutter_localizations/flutter_localizations.dart';
 void main() {
   Injector().initialiseAll([
     SessionDi(),
-    ChatDI(),
+    HubMessageDI(),
     NavigatorDI(),
     TalkDi(),
     UserDi(),
-    MessageDi(),
+    DirectMessageDi(),
     HomeDi(),
   ]);
   return runApp(MyApp(
